requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2; python_version >= "3.5"
chardet==3.0.4
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-bigquery==1.27.2
google-cloud-core==1.4.1
google-crc32c==1.0.0; python_version >= "3.5"
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20; python_version >= "3.5"
pytz==2020.1
requests==2.24.0
rsa==4.6; python_version >= "3.5"
six==1.15.0
urllib3==1.25.10
